# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /kg/species/{taxonKey}/context:
    get:
      operationId: getKgSpeciesContext
      tags: [kg]
      summary: Species graph context
      description: Returns n-hop neighbourhood subgraph for a species node
      parameters:
        - name: taxonKey
          in: path
          required: true
          schema:
            type: integer
        - name: hops
          in: query
          description: Number of hops to traverse (1 or 2, default 2)
          schema:
            type: integer
            default: 2
      responses:
        "200":
          description: Graph context
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/KgGraphContext"
        "404":
          description: Species not in graph
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /kg/nodes/{nodeId}/context:
    get:
      operationId: getKgNodeContext
      tags: [kg]
      summary: Generic node graph context
      description: Returns n-hop neighbourhood subgraph for any KG node by internal node id
      parameters:
        - name: nodeId
          in: path
          required: true
          schema:
            type: integer
        - name: hops
          in: query
          description: Number of hops to traverse (1 or 2, default 1)
          schema:
            type: integer
            default: 1
      responses:
        "200":
          description: Graph context
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/KgGraphContext"
        "404":
          description: Node not in graph
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /kg/insights/cooccurrence:
    get:
      operationId: getKgCooccurrence
      tags: [kg]
      summary: Co-occurrence clusters
      description: Returns species pairs that frequently co-occur (by Jaccard similarity)
      parameters:
        - name: region
          in: query
          description: Two-letter ISO country code to filter by region
          schema:
            type: string
        - name: minJaccard
          in: query
          description: Minimum Jaccard similarity threshold (default 0.1)
          schema:
            type: number
            default: 0.1
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
      responses:
        "200":
          description: Co-occurrence clusters
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CooccurrenceResponse"
  /kg/insights/endangered-hotspots:
    get:
      operationId: getKgEndangeredHotspots
      tags: [kg]
      summary: Endangered species hotspots
      description: Returns regions ranked by number of IUCN-threatened species
      parameters:
        - name: limit
          in: query
          schema:
            type: integer
            default: 10
      responses:
        "200":
          description: Endangered hotspots
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EndangeredHotspotsResponse"
  /kg/insights/taxonomy-gap:
    get:
      operationId: getKgTaxonomyGap
      tags: [kg]
      summary: Taxonomy data gaps
      description: Returns families and orders with fewest occurrences relative to species count
      parameters:
        - name: kingdom
          in: query
          schema:
            type: string
        - name: rank
          in: query
          schema:
            type: string
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
      responses:
        "200":
          description: Taxonomy gaps
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TaxonomyGapResponse"
  /kg/insights/threat-distribution:
    get:
      operationId: getKgThreatDistribution
      tags: [kg]
      summary: IUCN threat status distribution
      description: Returns IUCN status counts overall and by kingdom, plus families with highest threatened ratio
      parameters:
        - name: topFamilies
          in: query
          schema:
            type: integer
            default: 10
      responses:
        "200":
          description: Threat distribution
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ThreatDistributionResponse"
  /kg/insights/occurrence-trends:
    get:
      operationId: getKgOccurrenceTrends
      tags: [kg]
      summary: Occurrence trends over time
      description: Yearly occurrence counts plus per-country yearly series for top contributing countries
      parameters:
        - name: fromYear
          in: query
          schema:
            type: integer
            default: 1980
        - name: topCountries
          in: query
          schema:
            type: integer
            default: 5
      responses:
        "200":
          description: Occurrence trends
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OccurrenceTrendsResponse"
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
        message:
          type: string
      required:
        - error
        - message
...
    KgNode:
      type: object
      properties:
        nodeId:
          type: integer
        nodeType:
          type: string
        externalId:
          type: string
        label:
          type: string
        properties:
          type: object
          nullable: true
      required:
        - nodeId
        - nodeType
        - externalId
        - label
    KgEdge:
      type: object
      properties:
        edgeId:
          type: integer
        fromNode:
          type: integer
        toNode:
          type: integer
        edgeType:
          type: string
        weight:
          type: number
        properties:
          type: object
          nullable: true
      required:
        - edgeId
        - fromNode
        - toNode
        - edgeType
        - weight
    KgGraphContext:
      type: object
      properties:
        rootNode:
          $ref: "#/components/schemas/KgNode"
        hops:
          type: integer
        nodes:
          type: array
          items:
            $ref: "#/components/schemas/KgNode"
        edges:
          type: array
          items:
            $ref: "#/components/schemas/KgEdge"
      required:
        - rootNode
        - hops
        - nodes
        - edges
    CooccurrenceCluster:
      type: object
      properties:
        taxonKeyA:
          type: integer
        labelA:
          type: string
        taxonKeyB:
          type: integer
        labelB:
          type: string
        jaccardSimilarity:
          type: number
        sharedRegions:
          type: array
          items:
            type: string
      required:
        - taxonKeyA
        - labelA
        - taxonKeyB
        - labelB
        - jaccardSimilarity
        - sharedRegions
    CooccurrenceResponse:
      type: object
      properties:
        region:
          type: string
          nullable: true
        minJaccard:
          type: number
        count:
          type: integer
        clusters:
          type: array
          items:
            $ref: "#/components/schemas/CooccurrenceCluster"
      required:
        - region
        - minJaccard
        - count
        - clusters
    EndangeredSpecies:
      type: object
      properties:
        taxonKey:
          type: integer
        label:
          type: string
        iucnStatus:
          type: string
      required:
        - taxonKey
        - label
        - iucnStatus
    EndangeredHotspot:
      type: object
      properties:
        countryCode:
          type: string
        regionLabel:
          type: string
        endangeredCount:
          type: integer
        iucnStatuses:
          type: array
          items:
            type: string
        topSpecies:
          type: array
          items:
            $ref: "#/components/schemas/EndangeredSpecies"
      required:
        - countryCode
        - regionLabel
        - endangeredCount
        - iucnStatuses
        - topSpecies
    EndangeredHotspotsResponse:
      type: object
      properties:
        count:
          type: integer
        hotspots:
          type: array
          items:
            $ref: "#/components/schemas/EndangeredHotspot"
      required:
        - count
        - hotspots
    TaxonomyGapItem:
      type: object
      properties:
        rank:
          type: string
        taxonKey:
          type: integer
        label:
          type: string
        kingdom:
          type: string
          nullable: true
        totalSpeciesCount:
          type: integer
        occurrenceCount:
          type: integer
        occurrencePerSpecies:
          type: number
      required:
        - rank
        - taxonKey
        - label
        - totalSpeciesCount
        - occurrenceCount
        - occurrencePerSpecies
    TaxonomyGapResponse:
      type: object
      properties:
        count:
          type: integer
        gaps:
          type: array
          items:
            $ref: "#/components/schemas/TaxonomyGapItem"
      required:
        - count
        - gaps
    ThreatStatusBucket:
      type: object
      properties:
        status:
          type: string
        count:
          type: integer
      required: [status, count]
    KingdomThreatBreakdown:
      type: object
      properties:
        kingdom:
          type: string
        total:
          type: integer
        threatened:
          type: integer
        notEvaluated:
          type: integer
        byStatus:
          type: object
          additionalProperties:
            type: integer
      required: [kingdom, total, threatened, notEvaluated, byStatus]
    FamilyThreatItem:
      type: object
      properties:
        family:
          type: string
        kingdom:
          type: string
          nullable: true
        totalSpecies:
          type: integer
        threatenedSpecies:
          type: integer
        threatRatio:
          type: number
      required: [family, totalSpecies, threatenedSpecies, threatRatio]
    ThreatDistributionResponse:
      type: object
      properties:
        totalClassified:
          type: integer
        totalSpecies:
          type: integer
        byStatus:
          type: array
          items:
            $ref: "#/components/schemas/ThreatStatusBucket"
        byKingdom:
          type: array
          items:
            $ref: "#/components/schemas/KingdomThreatBreakdown"
        topThreatenedFamilies:
          type: array
          items:
            $ref: "#/components/schemas/FamilyThreatItem"
      required: [totalClassified, totalSpecies, byStatus, byKingdom, topThreatenedFamilies]
    YearOccurrencePoint:
      type: object
      properties:
        year:
          type: integer
        count:
          type: integer
      required: [year, count]
    CountryYearSeries:
      type: object
      properties:
        countryCode:
          type: string
        total:
          type: integer
        series:
          type: array
          items:
            $ref: "#/components/schemas/YearOccurrencePoint"
      required: [countryCode, total, series]
    OccurrenceTrendsResponse:
      type: object
      properties:
        yearMin:
          type: integer
        yearMax:
          type: integer
        totalWithYear:
          type: integer
        totalWithoutYear:
          type: integer
        yearly:
          type: array
          items:
            $ref: "#/components/schemas/YearOccurrencePoint"
        topCountries:
          type: array
          items:
            $ref: "#/components/schemas/CountryYearSeries"
      required: [yearMin, yearMax, totalWithYear, totalWithoutYear, yearly, topCountries]
...